% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{go_entailed_edges}
\alias{go_entailed_edges}
\title{retrieve table of entailed edges in GO (derivable from OWL base axioms)}
\usage{
go_entailed_edges(con, schema = "go")
}
\arguments{
\item{con}{live duckdb connection}

\item{schema}{character name, defaults to "go"}
}
\description{
retrieve table of entailed edges in GO (derivable from OWL base axioms)
}
