% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{go_statements}
\alias{go_statements}
\title{retrieve table of statements in GO}
\usage{
go_statements(con, schema = "go")
}
\arguments{
\item{con}{live duckdb connection}

\item{schema}{character name, defaults to "go"}
}
\description{
retrieve table of statements in GO
}
