% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_nodes.R
\name{go_terms}
\alias{go_terms}
\title{produce table with terms as a focus}
\usage{
go_terms(con, schema = "go", include_deprecated = FALSE)
}
\description{
produce table with terms as a focus
}
\examples{
con = make_go_con()
go_labels(con)
}
