% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connection.R
\name{make_go_con}
\alias{make_go_con}
\title{obtain a duckdb connection to GO SemanticSQL}
\usage{
make_go_con()
}
\description{
obtain a duckdb connection to GO SemanticSQL
}
\note{
uses ontoProc2 to obtain reference to SQLite reference to GO
}
